added grpc server stub and remove http

See merge request 7876010/go-8-middle-project!4

+++ cart/internal/model/cart.go
@@ -2,12 +2,12 @@ package model
 
 type Item struct {
 	SKU   SKU
-	Count uint16
+	Count uint32
 }
 
 type ItemDetail struct {
 	SKU   SKU
-	Count uint16
+	Count uint32
 	Price uint32
 	Name  string
 }
