+++ loms/internal/repository/repository.go
@@ -18,7 +18,7 @@ type Order interface {
 
 type Stock interface {
 	FindBySKU(ctx context.Context, sku model.SKU) (*model.Stock, error)
-	AddReserve(ctx context.Context, sku model.SKU, count uint16) error
-	CancelReserve(ctx context.Context, sku model.SKU, count uint16) error
-	ShipReserve(ctx context.Context, sku model.SKU, count uint16) error
+	AddReserve(ctx context.Context, sku model.SKU, count uint32) error
+	CancelReserve(ctx context.Context, sku model.SKU, count uint32) error
+	ShipReserve(ctx context.Context, sku model.SKU, count uint32) error
 }
